Crate ethcore_rpc

Ethcore rpc.

Modules

v1
Ethcore rpc v1.

Structs

ConfirmationsQueue
Queue for all unconfirmed requests.

IpcServer

NetworkSettings
Networking & RPC settings

RpcServer
Http server.

Server
jsonrpc http server instance

ServerBuilder
Convenient JSON-RPC HTTP Server builder.

SignerService
Manages communication with Signer crate

Enums

IpcServerError

RpcServerError
RPC Server startup error.

Traits

Extendable
An object that can be extended with IoDelegates

SigningQueue
A queue of transactions awaiting to be confirmed and signed.

Functions

is_major_importing
Check if client is during major sync or during block import.