Struct ethcore_rpc::RpcServer

pub struct RpcServer {
    // some fields omitted
}

Http server.

Methods

impl RpcServer

fn new() -> RpcServer

Construct new http server object.

fn start_http(&self, addr: &SocketAddr, cors_domains: Option<Vec<String>>, allowed_hosts: Option<Vec<String>>, panic_handler: Arc<PanicHandler>) -> Result<Server, RpcServerError>

Start http server asynchronously and returns result with Server handle on success or an error.

fn start_ipc(&self, addr: &str) -> Result<Server, Error>

Start ipc server asynchronously and returns result with Server handle on success or an error.