Trait ethcore_rpc::SigningQueue

pub trait SigningQueue: Send + Sync {
    fn add_request(&self,
               request: ConfirmationPayload)
               -> Result<ConfirmationPromise, QueueAddError>;
    fn request_rejected(&self, id: U256) -> Option<ConfirmationRequest>;
    fn request_confirmed(&self,
                     id: U256,
                     result: Result<ConfirmationResponse, Error>)
                     -> Option<ConfirmationRequest>;
    fn peek(&self, id: &U256) -> Option<ConfirmationRequest>;
    fn requests(&self) -> Vec<ConfirmationRequest>;
    fn len(&self) -> usize;
    fn is_empty(&self) -> bool;
}

A queue of transactions awaiting to be confirmed and signed.

Required Methods

fn add_request(&self,
request: ConfirmationPayload)
-> Result<ConfirmationPromise, QueueAddError>

Add new request to the queue. Returns a ConfirmationPromise that can be used to await for resolution of given request.

fn request_rejected(&self, id: U256) -> Option<ConfirmationRequest>

Removes a request from the queue. Notifies possible token holders that request was rejected.

fn request_confirmed(&self,
id: U256,
result: Result<ConfirmationResponse, Error>)
-> Option<ConfirmationRequest>

Removes a request from the queue. Notifies possible token holders that request was confirmed and given hash was assigned.

fn peek(&self, id: &U256) -> Option<ConfirmationRequest>

Returns a request if it is contained in the queue.

fn requests(&self) -> Vec<ConfirmationRequest>

Return copy of all the requests in the queue.

fn len(&self) -> usize

Returns number of requests awaiting confirmation.

fn is_empty(&self) -> bool

Returns true if there are no requests awaiting confirmation.