Module ethcore_rpc::v1

Ethcore rpc v1.

Compliant with ethereum rpc.

Reexports

pub use self::traits::{Web3, Eth, EthFilter, EthSigning, Net, Parity, ParityAccounts, ParitySet, ParitySigning, Signer, Personal, Traces, Rpc};

Modules

block_import
Block import analysis functions.

tests
RPC unit test moduleS

traits
Ethereum rpc interfaces.

types
Structures used in RPC communication

Structs

ConfirmationsQueue
Queue for all unconfirmed requests.

EthClient
Eth rpc implementation.

EthClientOptions
Eth RPC options

EthFilterClient
Eth filter rpc implementation.

NetClient
Net rpc implementation.

NetworkSettings
Networking & RPC settings

ParityAccountsClient
Account management (personal) rpc implementation.

ParityClient
Parity implementation.

ParitySetClient
Parity-specific rpc interface for operations altering the settings.

PersonalClient
Account management (personal) rpc implementation.

RpcClient
RPC generic methods implementation.

SignerClient
Transactions confirmation (personal) rpc implementation.

SignerService
Manages communication with Signer crate

SigningQueueClient
Implementation of functions that require signing when no trusted signer is used.

SigningUnsafeClient
Implementation of functions that require signing when no trusted signer is used.

TracesClient
Traces api implementation.

Web3Client
Web3 rpc implementation.

Traits

SigningQueue
A queue of transactions awaiting to be confirmed and signed.