Struct ethcore_rpc::v1::ConfirmationsQueue

pub struct ConfirmationsQueue {
    // some fields omitted
}

Queue for all unconfirmed transactions.

Methods

impl ConfirmationsQueue

fn start_listening<F>(&self, listener: F) -> Result<(), QueueError> where F: Fn(QueueEvent) -> ()

Blocks the thread and starts listening for notifications regarding all actions in the queue. For each event, listener callback will be invoked. This method can be used only once (only single consumer of events can exist).

fn finish(&self)

Notifies consumer that the communcation is over. No more events will be sent after this function is invoked.