Struct ethcore_rpc::v1::EthSigningQueueClient

pub struct EthSigningQueueClient<C, M> where C: MiningBlockChainClient, M: MinerService {
    // some fields omitted
}

Implementation of functions that require signing when no trusted signer is used.

Methods

impl<C, M> EthSigningQueueClient<C, M> where C: MiningBlockChainClient, M: MinerService

fn new(signer: &Arc<SignerService>, client: &Arc<C>, miner: &Arc<M>, accounts: &Arc<AccountProvider>) -> Self

Creates a new signing queue client given shared signing queue.