Trait ethcore_rpc::v1::traits::eth::Eth

pub trait Eth: Sized + Send + Sync + 'static {
    fn protocol_version(&self, _: Params) -> Result<Value, Error> { ... }
    fn syncing(&self, _: Params) -> Result<Value, Error> { ... }
    fn hashrate(&self, _: Params) -> Result<Value, Error> { ... }
    fn author(&self, _: Params) -> Result<Value, Error> { ... }
    fn is_mining(&self, _: Params) -> Result<Value, Error> { ... }
    fn gas_price(&self, _: Params) -> Result<Value, Error> { ... }
    fn accounts(&self, _: Params) -> Result<Value, Error> { ... }
    fn block_number(&self, _: Params) -> Result<Value, Error> { ... }
    fn balance(&self, _: Params) -> Result<Value, Error> { ... }
    fn storage_at(&self, _: Params) -> Result<Value, Error> { ... }
    fn block_by_hash(&self, _: Params) -> Result<Value, Error> { ... }
    fn block_by_number(&self, _: Params) -> Result<Value, Error> { ... }
    fn transaction_count(&self, _: Params) -> Result<Value, Error> { ... }
    fn block_transaction_count_by_hash(&self, _: Params) -> Result<Value, Error> { ... }
    fn block_transaction_count_by_number(&self, _: Params) -> Result<Value, Error> { ... }
    fn block_uncles_count_by_hash(&self, _: Params) -> Result<Value, Error> { ... }
    fn block_uncles_count_by_number(&self, _: Params) -> Result<Value, Error> { ... }
    fn code_at(&self, _: Params) -> Result<Value, Error> { ... }
    fn sign(&self, _: Params) -> Result<Value, Error> { ... }
    fn send_transaction(&self, _: Params) -> Result<Value, Error> { ... }
    fn send_raw_transaction(&self, _: Params) -> Result<Value, Error> { ... }
    fn call(&self, _: Params) -> Result<Value, Error> { ... }
    fn estimate_gas(&self, _: Params) -> Result<Value, Error> { ... }
    fn transaction_by_hash(&self, _: Params) -> Result<Value, Error> { ... }
    fn transaction_by_block_hash_and_index(&self, _: Params) -> Result<Value, Error> { ... }
    fn transaction_by_block_number_and_index(&self, _: Params) -> Result<Value, Error> { ... }
    fn transaction_receipt(&self, _: Params) -> Result<Value, Error> { ... }
    fn uncle_by_block_hash_and_index(&self, _: Params) -> Result<Value, Error> { ... }
    fn uncle_by_block_number_and_index(&self, _: Params) -> Result<Value, Error> { ... }
    fn compilers(&self, _: Params) -> Result<Value, Error> { ... }
    fn compile_lll(&self, _: Params) -> Result<Value, Error> { ... }
    fn compile_solidity(&self, _: Params) -> Result<Value, Error> { ... }
    fn compile_serpent(&self, _: Params) -> Result<Value, Error> { ... }
    fn logs(&self, _: Params) -> Result<Value, Error> { ... }
    fn work(&self, _: Params) -> Result<Value, Error> { ... }
    fn submit_work(&self, _: Params) -> Result<Value, Error> { ... }
    fn submit_hashrate(&self, _: Params) -> Result<Value, Error> { ... }
    fn to_delegate(self) -> IoDelegate<Self> { ... }
}

Eth rpc interface.

Provided Methods

fn protocol_version(&self, _: Params) -> Result<Value, Error>

Returns protocol version.

fn syncing(&self, _: Params) -> Result<Value, Error>

Returns an object with data about the sync status or false. (wtf?)

fn hashrate(&self, _: Params) -> Result<Value, Error>

Returns the number of hashes per second that the node is mining with.

fn author(&self, _: Params) -> Result<Value, Error>

Returns block author.

fn is_mining(&self, _: Params) -> Result<Value, Error>

Returns true if client is actively mining new blocks.

fn gas_price(&self, _: Params) -> Result<Value, Error>

Returns current gas_price.

fn accounts(&self, _: Params) -> Result<Value, Error>

Returns accounts list.

fn block_number(&self, _: Params) -> Result<Value, Error>

Returns highest block number.

fn balance(&self, _: Params) -> Result<Value, Error>

Returns balance of the given account.

fn storage_at(&self, _: Params) -> Result<Value, Error>

Returns content of the storage at given address.

fn block_by_hash(&self, _: Params) -> Result<Value, Error>

Returns block with given hash.

fn block_by_number(&self, _: Params) -> Result<Value, Error>

Returns block with given number.

fn transaction_count(&self, _: Params) -> Result<Value, Error>

Returns the number of transactions sent from given address at given time (block number).

fn block_transaction_count_by_hash(&self, _: Params) -> Result<Value, Error>

Returns the number of transactions in a block with given hash.

fn block_transaction_count_by_number(&self, _: Params) -> Result<Value, Error>

Returns the number of transactions in a block with given block number.

fn block_uncles_count_by_hash(&self, _: Params) -> Result<Value, Error>

Returns the number of uncles in a block with given hash.

fn block_uncles_count_by_number(&self, _: Params) -> Result<Value, Error>

Returns the number of uncles in a block with given block number.

fn code_at(&self, _: Params) -> Result<Value, Error>

Returns the code at given address at given time (block number).

fn sign(&self, _: Params) -> Result<Value, Error>

Signs the data with given address signature.

fn send_transaction(&self, _: Params) -> Result<Value, Error>

Sends transaction.

fn send_raw_transaction(&self, _: Params) -> Result<Value, Error>

Sends signed transaction.

fn call(&self, _: Params) -> Result<Value, Error>

Call contract.

fn estimate_gas(&self, _: Params) -> Result<Value, Error>

Estimate gas needed for execution of given contract.

fn transaction_by_hash(&self, _: Params) -> Result<Value, Error>

Get transaction by it's hash.

fn transaction_by_block_hash_and_index(&self, _: Params) -> Result<Value, Error>

Returns transaction at given block hash and index.

fn transaction_by_block_number_and_index(&self, _: Params) -> Result<Value, Error>

Returns transaction by given block number and index.

fn transaction_receipt(&self, _: Params) -> Result<Value, Error>

Returns transaction receipt.

fn uncle_by_block_hash_and_index(&self, _: Params) -> Result<Value, Error>

Returns an uncles at given block and index.

fn uncle_by_block_number_and_index(&self, _: Params) -> Result<Value, Error>

Returns an uncles at given block and index.

fn compilers(&self, _: Params) -> Result<Value, Error>

Returns available compilers.

fn compile_lll(&self, _: Params) -> Result<Value, Error>

Compiles lll code.

fn compile_solidity(&self, _: Params) -> Result<Value, Error>

Compiles solidity.

fn compile_serpent(&self, _: Params) -> Result<Value, Error>

Compiles serpent.

fn logs(&self, _: Params) -> Result<Value, Error>

Returns logs matching given filter object.

fn work(&self, _: Params) -> Result<Value, Error>

Returns the hash of the current block, the seedHash, and the boundary condition to be met.

fn submit_work(&self, _: Params) -> Result<Value, Error>

Used for submitting a proof-of-work solution.

fn submit_hashrate(&self, _: Params) -> Result<Value, Error>

Used for submitting mining hashrate.

fn to_delegate(self) -> IoDelegate<Self>

Should be used to convert object to io delegate.