Trait ethcore_rpc::v1::traits::eth_signing::EthSigning

pub trait EthSigning: Sized + Send + Sync + 'static {
    fn sign(&self, _: Params, _: Ready);
    fn post_sign(&self, _: Params) -> Result<Value, Error>;
    fn send_transaction(&self, _: Params, _: Ready);
    fn post_transaction(&self, _: Params) -> Result<Value, Error>;
    fn check_request(&self, _: Params) -> Result<Value, Error>;
    fn decrypt_message(&self, _: Params, _: Ready);

    fn to_delegate(self) -> IoDelegate<Self> { ... }
}

Signing methods implementation relying on unlocked accounts.

Required Methods

fn sign(&self, _: Params, _: Ready)

Signs the data with given address signature.

fn post_sign(&self, _: Params) -> Result<Value, Error>

Posts sign request asynchronously. Will return a confirmation ID for later use with check_transaction.

fn send_transaction(&self, _: Params, _: Ready)

Sends transaction; will block for 20s to try to return the transaction hash. If it cannot yet be signed, it will return a transaction ID for later use with check_transaction.

fn post_transaction(&self, _: Params) -> Result<Value, Error>

Posts transaction asynchronously. Will return a transaction ID for later use with check_transaction.

fn check_request(&self, _: Params) -> Result<Value, Error>

Checks the progress of a previously posted request (transaction/sign). Should be given a valid send_transaction ID. Returns the transaction hash, the zero hash (not yet available), or the signature, or an error.

fn decrypt_message(&self, _: Params, _: Ready)

Decrypt some ECIES-encrypted message. First parameter is the address with which it is encrypted, second is the ciphertext.

Provided Methods

fn to_delegate(self) -> IoDelegate<Self>

Should be used to convert object to io delegate.