Trait ethcore_rpc::v1::traits::parity::Parity

pub trait Parity: Sized + Send + Sync + 'static {
    fn accounts_info(&self,
                 Trailing<DappId>)
                 -> Result<BTreeMap<String, BTreeMap<String, String>>, Error>;
    fn transactions_limit(&self) -> Result<usize, Error>;
    fn extra_data(&self) -> Result<Bytes, Error>;
    fn gas_floor_target(&self) -> Result<U256, Error>;
    fn gas_ceil_target(&self) -> Result<U256, Error>;
    fn min_gas_price(&self) -> Result<U256, Error>;
    fn dev_logs(&self) -> Result<Vec<String>, Error>;
    fn dev_logs_levels(&self) -> Result<String, Error>;
    fn net_chain(&self) -> Result<String, Error>;
    fn net_peers(&self) -> Result<Peers, Error>;
    fn net_port(&self) -> Result<u16, Error>;
    fn rpc_settings(&self) -> Result<RpcSettings, Error>;
    fn node_name(&self) -> Result<String, Error>;
    fn default_extra_data(&self) -> Result<Bytes, Error>;
    fn gas_price_histogram(&self) -> Result<Histogram, Error>;
    fn unsigned_transactions_count(&self) -> Result<usize, Error>;
    fn generate_secret_phrase(&self) -> Result<String, Error>;
    fn phrase_to_address(&self, String) -> Result<H160, Error>;
    fn registry_address(&self) -> Result<Option<H160>, Error>;
    fn list_accounts(&self,
                 u64,
                 Option<H160>,
                 Trailing<BlockNumber>)
                 -> Result<Option<Vec<H160>>, Error>;
    fn list_storage_keys(&self,
                     H160,
                     u64,
                     Option<H256>,
                     Trailing<BlockNumber>)
                     -> Result<Option<Vec<H256>>, Error>;
    fn encrypt_message(&self, H512, Bytes) -> Result<Bytes, Error>;
    fn pending_transactions(&self) -> Result<Vec<Transaction>, Error>;
    fn future_transactions(&self) -> Result<Vec<Transaction>, Error>;
    fn pending_transactions_stats(&self)
                              -> Result<BTreeMap<H256, TransactionStats>, Error>;
    fn local_transactions(&self)
                      -> Result<BTreeMap<H256, LocalTransactionStatus>, Error>;
    fn signer_port(&self) -> Result<u16, Error>;
    fn dapps_port(&self) -> Result<u16, Error>;
    fn dapps_interface(&self) -> Result<String, Error>;
    fn next_nonce(&self, H160) -> Result<U256, Error>;
    fn mode(&self) -> Result<String, Error>;
    fn enode(&self) -> Result<String, Error>;
    fn consensus_capability(&self) -> Result<ConsensusCapability, Error>;
    fn version_info(&self) -> Result<VersionInfo, Error>;
    fn releases_info(&self) -> Result<Option<OperationsInfo>, Error>;
    fn chain_status(&self) -> Result<ChainStatus, Error>;

    fn to_delegate(self) -> IoDelegate<Self> { ... }
}

Parity-specific rpc interface.

Required Methods

fn accounts_info(&self,
Trailing<DappId>)
-> Result<BTreeMap<String, BTreeMap<String, String>>, Error>

Returns accounts information.

fn transactions_limit(&self) -> Result<usize, Error>

Returns current transactions limit.

fn extra_data(&self) -> Result<Bytes, Error>

Returns mining extra data.

fn gas_floor_target(&self) -> Result<U256, Error>

Returns mining gas floor target.

fn gas_ceil_target(&self) -> Result<U256, Error>

Returns mining gas floor cap.

fn min_gas_price(&self) -> Result<U256, Error>

Returns minimal gas price for transaction to be included in queue.

fn dev_logs(&self) -> Result<Vec<String>, Error>

Returns latest logs

fn dev_logs_levels(&self) -> Result<String, Error>

Returns logs levels

fn net_chain(&self) -> Result<String, Error>

Returns chain name

fn net_peers(&self) -> Result<Peers, Error>

Returns peers details

fn net_port(&self) -> Result<u16, Error>

Returns network port

fn rpc_settings(&self) -> Result<RpcSettings, Error>

Returns rpc settings

fn node_name(&self) -> Result<String, Error>

Returns node name

fn default_extra_data(&self) -> Result<Bytes, Error>

Returns default extra data

fn gas_price_histogram(&self) -> Result<Histogram, Error>

Returns distribution of gas price in latest blocks.

fn unsigned_transactions_count(&self) -> Result<usize, Error>

Returns number of unsigned transactions waiting in the signer queue (if signer enabled) Returns error when signer is disabled

fn generate_secret_phrase(&self) -> Result<String, Error>

Returns a cryptographically random phrase sufficient for securely seeding a secret key.

fn phrase_to_address(&self, String) -> Result<H160, Error>

Returns whatever address would be derived from the given phrase if it were to seed a brainwallet.

fn registry_address(&self) -> Result<Option<H160>, Error>

Returns the value of the registrar for this network.

fn list_accounts(&self,
u64,
Option<H160>,
Trailing<BlockNumber>)
-> Result<Option<Vec<H160>>, Error>

Returns all addresses if Fat DB is enabled (--fat-db), or null if not.

fn list_storage_keys(&self,
H160,
u64,
Option<H256>,
Trailing<BlockNumber>)
-> Result<Option<Vec<H256>>, Error>

Returns all storage keys of the given address (first parameter) if Fat DB is enabled (--fat-db), or null if not.

fn encrypt_message(&self, H512, Bytes) -> Result<Bytes, Error>

Encrypt some data with a public key under ECIES. First parameter is the 512-byte destination public key, second is the message.

fn pending_transactions(&self) -> Result<Vec<Transaction>, Error>

Returns all pending transactions from transaction queue.

fn future_transactions(&self) -> Result<Vec<Transaction>, Error>

Returns all future transactions from transaction queue.

fn pending_transactions_stats(&self)
-> Result<BTreeMap<H256, TransactionStats>, Error>

Returns propagation statistics on transactions pending in the queue.

fn local_transactions(&self)
-> Result<BTreeMap<H256, LocalTransactionStatus>, Error>

Returns a list of current and past local transactions with status details.

fn signer_port(&self) -> Result<u16, Error>

Returns current Trusted Signer port or an error if signer is disabled.

fn dapps_port(&self) -> Result<u16, Error>

Returns current Dapps Server port or an error if dapps server is disabled.

fn dapps_interface(&self) -> Result<String, Error>

Returns current Dapps Server interface address or an error if dapps server is disabled.

fn next_nonce(&self, H160) -> Result<U256, Error>

Returns next nonce for particular sender. Should include all transactions in the queue.

fn mode(&self) -> Result<String, Error>

Get the mode. Results one of: "active", "passive", "dark", "offline".

fn enode(&self) -> Result<String, Error>

Get the enode of this node.

fn consensus_capability(&self) -> Result<ConsensusCapability, Error>

Returns information on current consensus capability.

fn version_info(&self) -> Result<VersionInfo, Error>

Get our version information in a nice object.

fn releases_info(&self) -> Result<Option<OperationsInfo>, Error>

Get information concerning the latest releases if available.

fn chain_status(&self) -> Result<ChainStatus, Error>

Get the current chain status.

Provided Methods

fn to_delegate(self) -> IoDelegate<Self>

Transform this into an IoDelegate, automatically wrapping the parameters.