Trait ethcore_rpc::v1::traits::parity_set::ParitySet

pub trait ParitySet: Sized + Send + Sync + 'static {
    fn set_min_gas_price(&self, U256) -> Result<bool, Error>;
    fn set_gas_floor_target(&self, U256) -> Result<bool, Error>;
    fn set_gas_ceil_target(&self, U256) -> Result<bool, Error>;
    fn set_extra_data(&self, Bytes) -> Result<bool, Error>;
    fn set_author(&self, H160) -> Result<bool, Error>;
    fn set_engine_signer(&self, H160, String) -> Result<bool, Error>;
    fn set_transactions_limit(&self, usize) -> Result<bool, Error>;
    fn set_tx_gas_limit(&self, U256) -> Result<bool, Error>;
    fn add_reserved_peer(&self, String) -> Result<bool, Error>;
    fn remove_reserved_peer(&self, String) -> Result<bool, Error>;
    fn drop_non_reserved_peers(&self) -> Result<bool, Error>;
    fn accept_non_reserved_peers(&self) -> Result<bool, Error>;
    fn start_network(&self) -> Result<bool, Error>;
    fn stop_network(&self) -> Result<bool, Error>;
    fn set_mode(&self, String) -> Result<bool, Error>;
    fn hash_content(&self, Ready<H256>, String);
    fn upgrade_ready(&self) -> Result<Option<ReleaseInfo>, Error>;
    fn execute_upgrade(&self) -> Result<bool, Error>;

    fn to_delegate(self) -> IoDelegate<Self> { ... }
}

Parity-specific rpc interface for operations altering the settings.

Required Methods

fn set_min_gas_price(&self, U256) -> Result<bool, Error>

Sets new minimal gas price for mined blocks.

fn set_gas_floor_target(&self, U256) -> Result<bool, Error>

Sets new gas floor target for mined blocks.

fn set_gas_ceil_target(&self, U256) -> Result<bool, Error>

Sets new gas ceiling target for mined blocks.

fn set_extra_data(&self, Bytes) -> Result<bool, Error>

Sets new extra data for mined blocks.

fn set_author(&self, H160) -> Result<bool, Error>

Sets new author for mined block.

fn set_engine_signer(&self, H160, String) -> Result<bool, Error>

Sets account for signing consensus messages.

fn set_transactions_limit(&self, usize) -> Result<bool, Error>

Sets the limits for transaction queue.

fn set_tx_gas_limit(&self, U256) -> Result<bool, Error>

Sets the maximum amount of gas a single transaction may consume.

fn add_reserved_peer(&self, String) -> Result<bool, Error>

Add a reserved peer.

fn remove_reserved_peer(&self, String) -> Result<bool, Error>

Remove a reserved peer.

fn drop_non_reserved_peers(&self) -> Result<bool, Error>

Drop all non-reserved peers.

fn accept_non_reserved_peers(&self) -> Result<bool, Error>

Accept non-reserved peers (default behavior)

fn start_network(&self) -> Result<bool, Error>

Start the network.

Deprecated. Use set_mode("active") instead.

fn stop_network(&self) -> Result<bool, Error>

Stop the network.

Deprecated. Use set_mode("offline") instead.

fn set_mode(&self, String) -> Result<bool, Error>

Set the mode. Argument must be one of: "active", "passive", "dark", "offline".

fn hash_content(&self, Ready<H256>, String)

Hash a file content under given URL.

fn upgrade_ready(&self) -> Result<Option<ReleaseInfo>, Error>

Is there a release ready for install?

fn execute_upgrade(&self) -> Result<bool, Error>

Execute a release which is ready according to upgrade_ready().

Provided Methods

fn to_delegate(self) -> IoDelegate<Self>

Transform this into an IoDelegate, automatically wrapping the parameters.