Trait ethcore_rpc::v1::traits::parity_signing::ParitySigning

pub trait ParitySigning: Sized + Send + Sync + 'static {
    fn post_sign(&self,
             H160,
             H256)
             -> Result<Either<U256, ConfirmationResponse>, Error>;
    fn post_transaction(&self,
                    TransactionRequest)
                    -> Result<Either<U256, ConfirmationResponse>, Error>;
    fn check_request(&self, U256) -> Result<Option<ConfirmationResponse>, Error>;
    fn decrypt_message(&self, Ready<Bytes>, H160, Bytes);

    fn to_delegate(self) -> IoDelegate<Self> { ... }
}

Signing methods implementation.

Required Methods

fn post_sign(&self,
H160,
H256)
-> Result<Either<U256, ConfirmationResponse>, Error>

Posts sign request asynchronously. Will return a confirmation ID for later use with check_transaction.

fn post_transaction(&self,
TransactionRequest)
-> Result<Either<U256, ConfirmationResponse>, Error>

Posts transaction asynchronously. Will return a transaction ID for later use with check_transaction.

fn check_request(&self, U256) -> Result<Option<ConfirmationResponse>, Error>

Checks the progress of a previously posted request (transaction/sign). Should be given a valid send_transaction ID.

fn decrypt_message(&self, Ready<Bytes>, H160, Bytes)

Decrypt some ECIES-encrypted message. First parameter is the address with which it is encrypted, second is the ciphertext.

Provided Methods

fn to_delegate(self) -> IoDelegate<Self>

Transform this into an IoDelegate, automatically wrapping the parameters.