Trait ethcore_rpc::v1::traits::personal::PersonalAccounts

pub trait PersonalAccounts: Sized + Send + Sync + 'static {
    fn new_account(&self, String) -> Result<H160, Error>;
    fn new_account_from_phrase(&self, String, String) -> Result<H160, Error>;
    fn new_account_from_wallet(&self, String, String) -> Result<H160, Error>;
    fn new_account_from_secret(&self, H256, String) -> Result<H160, Error>;
    fn unlock_account(&self, H160, String, Option<u64>) -> Result<bool, Error>;
    fn test_password(&self, H160, String) -> Result<bool, Error>;
    fn change_password(&self, H160, String, String) -> Result<bool, Error>;
    fn sign_and_send_transaction(&self, TransactionRequest, String) -> Result<H256, Error>;
    fn set_account_name(&self, H160, String) -> Result<bool, Error>;
    fn set_account_meta(&self, H160, String) -> Result<bool, Error>;
    fn import_geth_accounts(&self, Vec<H160>) -> Result<Vec<H160>, Error>;
    fn geth_accounts(&self) -> Result<Vec<H160>, Error>;

    fn to_delegate(self) -> IoDelegate<Self> { ... }
}

Personal rpc methods altering stored accounts or their settings.

Required Methods

fn new_account(&self, String) -> Result<H160, Error>

Creates new account (it becomes new current unlocked account) Param is the password for the account.

fn new_account_from_phrase(&self, String, String) -> Result<H160, Error>

Creates new account from the given phrase using standard brainwallet mechanism. Second parameter is password for the new account.

fn new_account_from_wallet(&self, String, String) -> Result<H160, Error>

Creates new account from the given JSON wallet. Second parameter is password for the wallet and the new account.

fn new_account_from_secret(&self, H256, String) -> Result<H160, Error>

Creates new account from the given raw secret. Second parameter is password for the new account.

fn unlock_account(&self, H160, String, Option<u64>) -> Result<bool, Error>

Unlocks specified account for use (can only be one unlocked account at one moment)

fn test_password(&self, H160, String) -> Result<bool, Error>

Returns true if given password would unlock given account. Arguments: account, password.

fn change_password(&self, H160, String, String) -> Result<bool, Error>

Changes an account's password. Arguments: account, password, new_password.

fn sign_and_send_transaction(&self, TransactionRequest, String) -> Result<H256, Error>

Sends transaction and signs it in single call. The account is not unlocked in such case.

fn set_account_name(&self, H160, String) -> Result<bool, Error>

Set an account's name.

fn set_account_meta(&self, H160, String) -> Result<bool, Error>

Set an account's metadata string.

fn import_geth_accounts(&self, Vec<H160>) -> Result<Vec<H160>, Error>

Imports a number of Geth accounts, with the list provided as the argument.

fn geth_accounts(&self) -> Result<Vec<H160>, Error>

Returns the accounts available for importing from Geth.

Provided Methods

fn to_delegate(self) -> IoDelegate<Self>

Transform this into an IoDelegate, automatically wrapping the parameters.