Trait ethcore_rpc::v1::traits::traces::Traces

pub trait Traces: Sized + Send + Sync + 'static {
    fn filter(&self, _: Params) -> Result<Value, Error>;
    fn trace(&self, _: Params) -> Result<Value, Error>;
    fn transaction_traces(&self, _: Params) -> Result<Value, Error>;
    fn block_traces(&self, _: Params) -> Result<Value, Error>;
    fn call(&self, _: Params) -> Result<Value, Error>;
    fn raw_transaction(&self, _: Params) -> Result<Value, Error>;
    fn replay_transaction(&self, _: Params) -> Result<Value, Error>;

    fn to_delegate(self) -> IoDelegate<Self> { ... }
}

Traces specific rpc interface.

Required Methods

fn filter(&self, _: Params) -> Result<Value, Error>

Returns traces matching given filter.

fn trace(&self, _: Params) -> Result<Value, Error>

Returns transaction trace at given index.

fn transaction_traces(&self, _: Params) -> Result<Value, Error>

Returns all traces of given transaction.

fn block_traces(&self, _: Params) -> Result<Value, Error>

Returns all traces produced at given block.

fn call(&self, _: Params) -> Result<Value, Error>

Executes the given call and returns a number of possible traces for it.

fn raw_transaction(&self, _: Params) -> Result<Value, Error>

Executes the given raw transaction and returns a number of possible traces for it.

fn replay_transaction(&self, _: Params) -> Result<Value, Error>

Executes the transaction with the given hash and returns a number of possible traces for it.

Provided Methods

fn to_delegate(self) -> IoDelegate<Self>

Should be used to convert object to io delegate.