Module ethcore_rpc::v1::types

Structures used in RPC communication

Structs

Block
Block representation

Bytes
Wrapper structure around vector of bytes.

CallRequest
Call request

ChainStatus
Chain status.

ConfirmationRequest
Confirmation waiting in a queue

ConfirmationResponseWithToken
Confirmation response with additional token for further requests

DappId
Dapplication Internal Id

DecryptRequest
Decrypt request

Filter
Filter

H160
Hash serialization

H2048
Hash serialization

H256
Hash serialization

H512
Hash serialization

H520
Hash serialization

H64
Hash serialization

Histogram
Values of RPC settings.

Index
Represents usize.

LocalizedTrace
Trace

Log
Log

OperationsInfo
Information on our operations environment.

PeerEthereumProtocolInfo
Peer Ethereum protocol information

PeerInfo
Peer connection information

PeerNetworkInfo
Peer network information

PeerProtocolsInfo
Peer protocols information

Peers
Peers info

Receipt
Receipt

ReleaseInfo
Information regarding a particular release of Parity

RichBlock
Block representation with additional info

RichRawTransaction
Geth-compatible output for eth_signTransaction method

RpcSettings
Values of RPC settings.

SignRequest
Sign request

SyncInfo
Sync info

TraceFilter
Trace filter

TraceResults
A diff of some chunk of memory.

Transaction
Transaction

TransactionModification
Possible modifications to the confirmed transaction sent by Trusted Signer

TransactionRequest
Transaction request coming from RPC

TransactionStats
Propagation statistics for pending transaction.

U128
Uint serialization.

U256
Uint serialization.

Version
Semantic version.

VersionInfo
Version information of a particular release.

Work
The result of an eth_getWork call: it differs based on an option whether to send the block number.

Enums

BlockNumber
Represents rpc api block number param.

BlockTransactions
Block Transactions

ConfirmationPayload
Confirmation payload, i.e. the thing to be confirmed

ConfirmationResponse
Confirmation response for particular payload

ConsensusCapability
Capability info

Either
Represents two possible return values.

FilterChanges
Results of the filter_changes RPC.

LocalTransactionStatus
Local Transaction Status

ReleaseTrack
A release's track.

SyncStatus
Sync status