Trait ethcore_rpc::v1::traits::ethcore::Ethcore

pub trait Ethcore: Sized + Send + Sync + 'static {
    fn transactions_limit(&self, _: Params) -> Result<Value, Error>;
    fn extra_data(&self, _: Params) -> Result<Value, Error>;
    fn gas_floor_target(&self, _: Params) -> Result<Value, Error>;
    fn gas_ceil_target(&self, _: Params) -> Result<Value, Error>;
    fn min_gas_price(&self, _: Params) -> Result<Value, Error>;
    fn dev_logs(&self, _: Params) -> Result<Value, Error>;
    fn dev_logs_levels(&self, _: Params) -> Result<Value, Error>;
    fn net_chain(&self, _: Params) -> Result<Value, Error>;
    fn net_max_peers(&self, _: Params) -> Result<Value, Error>;
    fn net_port(&self, _: Params) -> Result<Value, Error>;
    fn rpc_settings(&self, _: Params) -> Result<Value, Error>;
    fn node_name(&self, _: Params) -> Result<Value, Error>;
    fn default_extra_data(&self, _: Params) -> Result<Value, Error>;
    fn gas_price_statistics(&self, _: Params) -> Result<Value, Error>;
    fn unsigned_transactions_count(&self, _: Params) -> Result<Value, Error>;

    fn to_delegate(self) -> IoDelegate<Self> { ... }
}

Ethcore-specific rpc interface.

Required Methods

fn transactions_limit(&self, _: Params) -> Result<Value, Error>

Returns current transactions limit.

fn extra_data(&self, _: Params) -> Result<Value, Error>

Returns mining extra data.

fn gas_floor_target(&self, _: Params) -> Result<Value, Error>

Returns mining gas floor target.

fn gas_ceil_target(&self, _: Params) -> Result<Value, Error>

Returns mining gas floor cap.

fn min_gas_price(&self, _: Params) -> Result<Value, Error>

Returns minimal gas price for transaction to be included in queue.

fn dev_logs(&self, _: Params) -> Result<Value, Error>

Returns latest logs

fn dev_logs_levels(&self, _: Params) -> Result<Value, Error>

Returns logs levels

fn net_chain(&self, _: Params) -> Result<Value, Error>

Returns chain name

fn net_max_peers(&self, _: Params) -> Result<Value, Error>

Returns max peers

fn net_port(&self, _: Params) -> Result<Value, Error>

Returns network port

fn rpc_settings(&self, _: Params) -> Result<Value, Error>

Returns rpc settings

fn node_name(&self, _: Params) -> Result<Value, Error>

Returns node name

fn default_extra_data(&self, _: Params) -> Result<Value, Error>

Returns default extra data

fn gas_price_statistics(&self, _: Params) -> Result<Value, Error>

Returns distribution of gas price in latest blocks.

fn unsigned_transactions_count(&self, _: Params) -> Result<Value, Error>

Returns number of unsigned transactions waiting in the signer queue (if signer enabled) Returns error when signer is disabled

Provided Methods

fn to_delegate(self) -> IoDelegate<Self>

Should be used to convert object to io delegate.