Trait ethcore_rpc::v1::traits::traces::Traces

pub trait Traces: Sized + Send + Sync + 'static {
    fn filter(&self, TraceFilter) -> Result<Vec<LocalizedTrace>, Error>;
    fn trace(&self, H256, Vec<Index>) -> Result<Option<LocalizedTrace>, Error>;
    fn transaction_traces(&self, H256) -> Result<Vec<LocalizedTrace>, Error>;
    fn block_traces(&self, BlockNumber) -> Result<Vec<LocalizedTrace>, Error>;
    fn call(&self,
        CallRequest,
        Vec<String>,
        Trailing<BlockNumber>)
        -> Result<Option<TraceResults>, Error>;
    fn raw_transaction(&self,
                   Bytes,
                   Vec<String>,
                   Trailing<BlockNumber>)
                   -> Result<Option<TraceResults>, Error>;
    fn replay_transaction(&self,
                      H256,
                      Vec<String>)
                      -> Result<Option<TraceResults>, Error>;

    fn to_delegate(self) -> IoDelegate<Self> { ... }
}

Traces specific rpc interface.

Required Methods

fn filter(&self, TraceFilter) -> Result<Vec<LocalizedTrace>, Error>

Returns traces matching given filter.

fn trace(&self, H256, Vec<Index>) -> Result<Option<LocalizedTrace>, Error>

Returns transaction trace at given index.

fn transaction_traces(&self, H256) -> Result<Vec<LocalizedTrace>, Error>

Returns all traces of given transaction.

fn block_traces(&self, BlockNumber) -> Result<Vec<LocalizedTrace>, Error>

Returns all traces produced at given block.

fn call(&self,
CallRequest,
Vec<String>,
Trailing<BlockNumber>)
-> Result<Option<TraceResults>, Error>

Executes the given call and returns a number of possible traces for it.

fn raw_transaction(&self,
Bytes,
Vec<String>,
Trailing<BlockNumber>)
-> Result<Option<TraceResults>, Error>

Executes the given raw transaction and returns a number of possible traces for it.

fn replay_transaction(&self,
H256,
Vec<String>)
-> Result<Option<TraceResults>, Error>

Executes the transaction with the given hash and returns a number of possible traces for it.

Provided Methods

fn to_delegate(self) -> IoDelegate<Self>

Transform this into an IoDelegate, automatically wrapping the parameters.